requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
build==1.4.2
    # via
    #   -r requirements.in
    #   pip-tools
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   pip-tools
    #   uvicorn
fastapi==0.135.3
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
packaging==26.0
    # via
    #   build
    #   wheel
pillow==12.2.0
    # via -r requirements.in
pip-tools==7.5.3
    # via -r requirements.in
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.24
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   uvicorn
requests==2.33.1
    # via -r requirements.in
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.44.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via
    #   -r requirements.in
    #   uvicorn
websockets==16.0
    # via uvicorn
wheel==0.46.3
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
